requirements.txt
asgiref==3.5.0
certifi==2021.10.8
Django==3.2.11
django-cors-headers==3.11.0
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.2
django-nine==0.2.5
djangorestframework==3.13.1
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
ipaddress==1.0.23
# pkg_resources==0.0.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
